+++ gitnexus/test/unit/cfg/java-visitor.test.ts
@@ -2,7 +2,17 @@ import { describe, it, expect, vi } from 'vitest';
 import { createRequire } from 'node:module';
 import { createJavaCfgVisitor } from '../../../src/core/ingestion/cfg/visitors/java.js';
 import type { FunctionCfg, SiteRecord } from '../../../src/core/ingestion/cfg/types.js';
-import { makeCfgHarness, type CfgHarness } from '../../helpers/cfg-harness.js';
+import {
+  makeCfgHarness,
+  type CfgHarness,
+  block,
+  edgeKinds,
+  reaches,
+  reachable,
+  bindingIdx,
+  allSites,
+  hasAnySites,
+} from '../../helpers/cfg-harness.js';
 import { isExitReachableFromAllBlocks } from '../../../src/core/ingestion/cfg/post-dominators.js';
 import { computeControlDependence } from '../../../src/core/ingestion/cfg/control-dependence.js';
 
@@ -16,54 +26,6 @@ const javaGrammar = createRequire(import.meta.url)('tree-sitter-java') as Parame
 >[0];
 
 const java: CfgHarness = makeCfgHarness(javaGrammar, createJavaCfgVisitor(), 'fixture.java');
-
-const block = (cfg: FunctionCfg, substr: string): number => {
-  const b = cfg.blocks.find((bl) => bl.text.includes(substr));
-  if (!b) throw new Error(`no block containing ${JSON.stringify(substr)}`);
-  return b.index;
-};
-
-const edgeKinds = (cfg: FunctionCfg): Set<string> => new Set(cfg.edges.map((e) => e.kind));
-
-function reaches(cfg: FunctionCfg, from: number, to: number): boolean {
-  const adj = new Map<number, number[]>();
-  for (const e of cfg.edges) (adj.get(e.from) ?? adj.set(e.from, []).get(e.from)!).push(e.to);
-  const seen = new Set([from]);
-  const stack = [from];
-  while (stack.length) {
-    const n = stack.pop() as number;
-    if (n === to) return true;
-    for (const nx of adj.get(n) ?? []) if (!seen.has(nx)) (seen.add(nx), stack.push(nx));
-  }
-  return seen.has(to);
-}
-const reachable = (cfg: FunctionCfg, idx: number): boolean => reaches(cfg, cfg.entryIndex, idx);
-
-/** Is EXIT reverse-reachable from every reachable block? (CDG soundness gate.) */
-function exitReachableFromAll(cfg: FunctionCfg): boolean {
-  for (const b of cfg.blocks) {
-    if (b.index === cfg.exitIndex) continue;
-    if (!reachable(cfg, b.index)) continue; // unreachable blocks exempt
-    if (!reaches(cfg, b.index, cfg.exitIndex)) return false;
-  }
-  return true;
-}
-
-/** Resolve a binding by name → its index in the function's binding table. */
-function bindingIdx(cfg: FunctionCfg, name: string): number {
-  const i = (cfg.bindings ?? []).findIndex((b) => b.name === name);
-  if (i < 0) throw new Error(`no binding ${name}`);
-  return i;
-}
-
-/** Every taint `SiteRecord` harvested across the function's statements. */
-function allSites(cfg: FunctionCfg): SiteRecord[] {
-  const out: SiteRecord[] = [];
-  for (const b of cfg.blocks) for (const s of b.statements ?? []) out.push(...(s.sites ?? []));
-  return out;
-}
-const hasAnySites = (cfg: FunctionCfg): boolean =>
-  cfg.blocks.some((b) => (b.statements ?? []).some((s) => (s.sites ?? []).length > 0));
 
 const hasDef = (cfg: FunctionCfg, idx: number): boolean =>
   cfg.blocks.some((bl) => bl.statements?.some((s) => s.defs.includes(idx)));
