+++ gitnexus/test/unit/cfg/swift-visitor.test.ts
@@ -2,7 +2,15 @@ import { describe, it, expect } from 'vitest';
 import { requireVendoredGrammar } from '../../../src/core/tree-sitter/vendored-grammars.js';
 import { createSwiftCfgVisitor } from '../../../src/core/ingestion/cfg/visitors/swift.js';
 import type { FunctionCfg } from '../../../src/core/ingestion/cfg/types.js';
-import { makeCfgHarness, type CfgHarness } from '../../helpers/cfg-harness.js';
+import {
+  makeCfgHarness,
+  type CfgHarness,
+  block,
+  edgeKinds,
+  reaches,
+  reachable,
+  bindingIdx,
+} from '../../helpers/cfg-harness.js';
 import { isExitReachableFromAllBlocks } from '../../../src/core/ingestion/cfg/post-dominators.js';
 import { computeControlDependence } from '../../../src/core/ingestion/cfg/control-dependence.js';
 
@@ -18,45 +26,6 @@ const swiftGrammar = requireVendoredGrammar('tree-sitter-swift') as Parameters<
 >[0];
 
 const swift: CfgHarness = makeCfgHarness(swiftGrammar, createSwiftCfgVisitor(), 'fixture.swift');
-
-const block = (cfg: FunctionCfg, substr: string): number => {
-  const b = cfg.blocks.find((bl) => bl.text.includes(substr));
-  if (!b) throw new Error(`no block containing ${JSON.stringify(substr)}`);
-  return b.index;
-};
-
-const edgeKinds = (cfg: FunctionCfg): Set<string> => new Set(cfg.edges.map((e) => e.kind));
-
-function reaches(cfg: FunctionCfg, from: number, to: number): boolean {
-  const adj = new Map<number, number[]>();
-  for (const e of cfg.edges) (adj.get(e.from) ?? adj.set(e.from, []).get(e.from)!).push(e.to);
-  const seen = new Set([from]);
-  const stack = [from];
-  while (stack.length) {
-    const n = stack.pop() as number;
-    if (n === to) return true;
-    for (const nx of adj.get(n) ?? []) if (!seen.has(nx)) (seen.add(nx), stack.push(nx));
-  }
-  return seen.has(to);
-}
-const reachable = (cfg: FunctionCfg, idx: number): boolean => reaches(cfg, cfg.entryIndex, idx);
-
-/** Is EXIT reverse-reachable from every reachable block? (CDG soundness gate.) */
-function exitReachableFromAll(cfg: FunctionCfg): boolean {
-  for (const b of cfg.blocks) {
-    if (b.index === cfg.exitIndex) continue;
-    if (!reachable(cfg, b.index)) continue; // unreachable blocks exempt
-    if (!reaches(cfg, b.index, cfg.exitIndex)) return false;
-  }
-  return true;
-}
-
-/** Resolve a binding by name → its index in the function's binding table. */
-function bindingIdx(cfg: FunctionCfg, name: string): number {
-  const i = (cfg.bindings ?? []).findIndex((b) => b.name === name);
-  if (i < 0) throw new Error(`no binding ${name}`);
-  return i;
-}
 
 describe('Swift CfgVisitor — structure', () => {
   it('straight-line body: ENTRY → block → EXIT (seq)', () => {
@@ -187,7 +156,7 @@ describe('Swift CfgVisitor — loops', () => {
   it('repeat {} while true keeps EXIT reverse-reachable', () => {
     const cfg = swift.cfgOf(`func f() { repeat { work() } while true }`);
     expect(edgeKinds(cfg).has('cond-false')).toBe(true);
-    expect(exitReachableFromAll(cfg)).toBe(true);
+    expect(isExitReachableFromAllBlocks(cfg)).toBe(true);
     expect(reaches(cfg, cfg.entryIndex, cfg.exitIndex)).toBe(true);
   });
 });
