+++ gitnexus/test/unit/cfg/typescript-visitor.test.ts
@@ -7,6 +7,7 @@ import {
   TS_FUNCTION_TYPES,
 } from '../../../src/core/ingestion/cfg/visitors/typescript.js';
 import type { FunctionCfg } from '../../../src/core/ingestion/cfg/types.js';
+import { block, edgeKinds, reaches, reachable } from '../../helpers/cfg-harness.js';
 
 // U2 — the TS/JS CfgVisitor, one hazard per test. Each fixture's distinctive
 // statement text (markerWork(), handleErr(), cleanup(), …) lets us find the
@@ -45,30 +46,6 @@ function cfgOf(code: string, index = 0): FunctionCfg {
   if (!cfg) throw new Error('buildFunctionCfg returned undefined');
   return cfg;
 }
-
-const block = (cfg: FunctionCfg, substr: string): number => {
-  const b = cfg.blocks.find((bl) => bl.text.includes(substr));
-  if (!b) throw new Error(`no block containing ${JSON.stringify(substr)}`);
-  return b.index;
-};
-
-const edgeKinds = (cfg: FunctionCfg): Set<string> => new Set(cfg.edges.map((e) => e.kind));
-
-/** Does control reach `to` from `from` following edges? */
-function reaches(cfg: FunctionCfg, from: number, to: number): boolean {
-  const adj = new Map<number, number[]>();
-  for (const e of cfg.edges) (adj.get(e.from) ?? adj.set(e.from, []).get(e.from)!).push(e.to);
-  const seen = new Set([from]);
-  const stack = [from];
-  while (stack.length) {
-    const n = stack.pop() as number;
-    if (n === to) return true;
-    for (const nx of adj.get(n) ?? []) if (!seen.has(nx)) (seen.add(nx), stack.push(nx));
-  }
-  return seen.has(to);
-}
-
-const reachable = (cfg: FunctionCfg, idx: number): boolean => reaches(cfg, cfg.entryIndex, idx);
 
 describe('TS/JS CfgVisitor — structure', () => {
   it('straight-line body: ENTRY → block → EXIT', () => {
